+++ src/involvement/views/view_position.py
@@ -19,43 +19,48 @@ def view_position(request, context, page, position=None):
 
     # Load application form if user is logged in
     if request.user.is_authenticated:
-        # Did the user already have an application?
-        try:
-            appl = Application.objects.get(applicant=request.user,
-                                           position=context['position'])
-            context['status'] = appl.status
-        except ObjectDoesNotExist:
-            appl = Application()
-            context['status'] = 'draft'
-        # Did the user already fill in the form?
-        if request.method == 'POST' and not context['position'].is_past_due:
-            context['form'] = ApplicationForm(request.POST,
-                                              instance=appl)
-            context['reference_forms'] = ReferenceFormSet(request.POST,
-                                                          request.FILES,
-                                                          instance=appl)
-            if context['form'].is_valid() \
-                    and context['reference_forms'].is_valid():
-                appl = context['form'].save(commit=False)
-                appl.applicant = request.user
-                appl.position = context['position']
-                appl.save()
-                context['reference_forms'].save()
-                if not appl.status == 'draft':
-                    messages.add_message(
-                        request,
-                        messages.SUCCESS,
-                        _('Your application has been submitted!'),
-                    )
-                    return HttpResponseRedirect(
-                        page.get_url(request=request)
-                        + page.reverse_subpage('my_applications')
-                    )
-            else:
-                return render(request, 'involvement/position.html', context)
+        if request.user.melos_id:
+            context['membership_status'] = request.user.get_status
+            context['email'] = request.user.get_email
+            # Did the user already have an application?
+            try:
+                appl = Application.objects.get(applicant=request.user,
+                                               position=context['position'])
+                context['status'] = appl.status
+            except ObjectDoesNotExist:
+                appl = Application()
+                context['status'] = 'draft'
+            # Did the user already fill in the form?
+            if (request.method == 'POST'
+               and not context['position'].is_past_due):
+                context['form'] = ApplicationForm(request.POST,
+                                                  instance=appl)
+                context['reference_forms'] = ReferenceFormSet(request.POST,
+                                                              request.FILES,
+                                                              instance=appl)
+                if context['form'].is_valid() \
+                        and context['reference_forms'].is_valid():
+                    appl = context['form'].save(commit=False)
+                    appl.applicant = request.user
+                    appl.position = context['position']
+                    appl.save()
+                    context['reference_forms'].save()
+                    if not appl.status == 'draft':
+                        messages.add_message(
+                            request,
+                            messages.SUCCESS,
+                            _('Your application has been submitted!'),
+                        )
+                        return HttpResponseRedirect(
+                            page.get_url(request=request)
+                            + page.reverse_subpage('my_applications')
+                        )
+                else:
+                    return render(request, 'involvement/position.html',
+                                  context)
 
-        # Render fresh: empty or after saving draft.
-        context['form'] = ApplicationForm(instance=appl)
-        context['reference_forms'] = ReferenceFormSet(instance=appl)
+            # Render fresh: empty or after saving draft.
+            context['form'] = ApplicationForm(instance=appl)
+            context['reference_forms'] = ReferenceFormSet(instance=appl)
 
     return render(request, 'involvement/position.html', context)