+++ src/members/models/member.py
@@ -1,34 +1,70 @@
-import requests
-from datetime import date
+from datetime import date, timedelta
 from django.apps import apps
-from django.conf import settings
 from django.contrib.auth.models import AbstractUser, UserManager
 from django.core import validators
 from django.db import models
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
-from requests.auth import HTTPDigestAuth
 from simple_email_confirmation.models import SimpleEmailConfirmationUserMixin
+from utils.melos_client import MelosClient
+from utils.validators import SSNValidator
+
+
+def status_changed_default():
+    return timezone.now() - timedelta(days=2)
 
 
 class CaseInsensitiveUsernameUserManager(UserManager):
-    # Get username by insensitive case
+    # Search username with insensitive case
     def get_by_natural_key(self, username):
         case_insensitive_username_field = '{}__iexact' \
             .format(self.model.USERNAME_FIELD)
         return self.get(**{case_insensitive_username_field: username})
 
 
+class MelosUserManager(CaseInsensitiveUsernameUserManager):
+    # Search Member through Melos if username is SSN
+    def get_by_natural_key(self, username):
+        member = None
+        try:
+            member = super().get_by_natural_key(username)
+        except Exception as e:
+            member = Member.find_by_ssn(username)
+
+            if member is None:
+                raise e
+
+        return member
+
+
 class Member(SimpleEmailConfirmationUserMixin, AbstractUser):
     """This class describes a member"""
 
-    objects = CaseInsensitiveUsernameUserManager()
+    # ---- AbstractUser overrides ---
+
+    objects = MelosUserManager()
+
+    first_name = models.CharField(
+        verbose_name=_('first name'),
+        max_length=30,
+        blank=True,
+        null=True
+    )
+
+    last_name = models.CharField(
+        verbose_name=_('last name'),
+        max_length=150,
+        blank=True,
+        null=True
+    )
 
     # ---- Personal information ------
 
     birthday = models.DateField(
         verbose_name=_('Birthday'),
-        null=True
+        blank=True,
+        null=True,
+        editable=False
     )
 
     person_number_ext = models.CharField(
@@ -42,6 +78,7 @@ class Member(SimpleEmailConfirmationUserMixin, AbstractUser):
         )],
         unique_for_date="birthday",
         blank=True,
+        null=True
     )
 
     # ---- Membership information ------
@@ -60,8 +97,9 @@ class Member(SimpleEmailConfirmationUserMixin, AbstractUser):
         blank=False,
         default='unknown'
     )
+
     status_changed = models.DateTimeField(
-        default=timezone.now,
+        default=status_changed_default,
         null=False,
     )
 
@@ -71,11 +109,8 @@ class Member(SimpleEmailConfirmationUserMixin, AbstractUser):
         max_length=20,
         verbose_name=_('Phone number'),
         help_text=_('Enter a phone number so UTN may reach you'),
-        validators=[validators.RegexValidator(
-            regex=r'^\+?\d+$',
-            message=_('Please enter a valid phone number'),
-        )],
         blank=True,
+        default='',
     )
 
     # ---- University information ------
@@ -108,11 +143,19 @@ class Member(SimpleEmailConfirmationUserMixin, AbstractUser):
         blank=True,
     )
 
+    # ---- Melos ------
+
+    melos_id = models.IntegerField(
+        blank=True,
+        editable=False,
+        null=True,
+        unique=True,
+    )
+
+    melos_user_data = None
+
     def __str__(self) -> str:
-        if self.first_name and self.last_name:
-            return self.get_full_name()
-        else:
-            return self.username
+        return self.username
 
     @property
     def teams(self):
@@ -134,42 +177,67 @@ class Member(SimpleEmailConfirmationUserMixin, AbstractUser):
             positions__applications__status='appointed',
         )
 
-    def person_number(self) -> str:
-        if self.birthday is None or self.person_number_ext is None:
-            return ''
-        else:
-            return '%s-%s' % (self.birthday.strftime('%Y%m%d'),
-                              self.person_number_ext)
+    @property
+    def get_status(self):
+        self.update_status()
+        return self.status
 
-    def update_status(self, data=None):
+    @property
+    def get_first_name(self):
+        data = self.get_melos_user_data()
+        return '' if data is None else data['first_name']
+
+    @property
+    def get_last_name(self):
+        data = self.get_melos_user_data()
+        return '' if data is None else data['last_name']
+
+    @property
+    def get_full_name(self):
+        return '{} {}'.format(self.get_first_name, self.get_last_name)
+
+    @property
+    def get_ssn(self):
+        data = self.get_melos_user_data()
+        return None if data is None else data['person_number']
+
+    @property
+    def get_email(self):
+        if not self.email:
+            data = self.get_melos_user_data()
+            if data | data['email']:
+                self.email = data['email']
+                self.save()
+
+        return self.email
+
+    def person_number(self) -> str:
+        return self.get_ssn
+
+    def update_status(self, data=None, save=True):
         if data is None:
-            if self.person_number() == '':
+            # Prevent updating this value to often
+            if timezone.now() - self.status_changed < timedelta(days=1):
                 return
-            try:
-                r = requests.get(
-                    'https://register.utn.se/api.php',
-                    auth=HTTPDigestAuth(settings.MEMBERSHIP_API_USER,
-                                        settings.MEMBERSHIP_API_PASSWORD),
-                    params={
-                        'action': 'check',
-                        'person_number': self.person_number().replace('-', '')
-                    },
-                )
-                data = r.json().get('status')
-            except requests.exceptions.ConnectionError:
-                data = 'unknown'
-            except ValueError:
+
+            melos_user_data = self.get_melos_user_data()
+            if melos_user_data is None:
                 return
+            is_member = MelosClient.is_member(melos_user_data['person_number'])
+            data = "member" if is_member else "nonmember"
 
         if data == 'member':
             self.status = 'member'
         elif data == 'nonmember':
-            if self.status in ['unknown', 'nonmember']:
+            if self.status not in ['member', 'alumnus']:
                 self.status = 'nonmember'
             else:
                 self.status = 'alumnus'
 
         self.status_changed = timezone.now()
+
+        if save:
+            self.save()
 
     def remove_old_email(self):
         for email in self.get_unconfirmed_emails() or []:
@@ -177,3 +245,46 @@ class Member(SimpleEmailConfirmationUserMixin, AbstractUser):
         for email in self.get_confirmed_emails():
             if email != self.email:
                 self.remove_email(email)
+
+    def get_status_text(self, status):
+        status = _('Member') if status == 'member' else _('Nonmember')
+        return status
+
+    def sync_user_groups(self):
+        current_groups = self.groups.all()
+        wanted_groups = list(map(lambda role: role.group, self.roles.filter(
+            positions__applications__removed=False
+        ).all()))
+
+        # Remove unavailable groups
+        for group in current_groups:
+            if group not in wanted_groups:
+                group.user_set.remove(self)
+
+        # Add missing groups
+        for group in wanted_groups:
+            if (group not in current_groups):
+                group.user_set.add(self)
+
+    def get_melos_user_data(self):
+        if self.melos_user_data is None:
+            self.melos_user_data = MelosClient.get_user_data(self.melos_id)
+        return self.melos_user_data
+
+    @staticmethod
+    def find_by_melos_id(melos_id):
+        if melos_id:
+            return Member.objects.filter(melos_id=int(melos_id)).first()
+        return None
+
+    @staticmethod
+    def find_by_ssn(ssn):
+        try:
+            ssn = ssn.strip()
+            SSNValidator()(ssn)
+            melos_id = MelosClient.get_melos_id(ssn)
+            return Member.find_by_melos_id(melos_id)
+        except Exception:
+            pass
+
+        return None